fix: use app_startup and app_shutdown instead of lifespan
The lifespan cleanup runs before app shutdown so the registry is already closed. If users neeed to access the registry and get a service for any cleanup as part of the app shutdown, they won't have access to the registry. Instead, ensure that the registry is created and attached to the app first on startup, and that it is the last function ran on app shutdown.